!fix(Client): Do not hold goal_handle ptr if result cb is set.
The capture of the lambda held an copy of the shared_ptr to the
goal handle. This lead to the behavior that the goal callbacks
would be called, even though the user dropped the handle.

Note, this will break the current tutorial, as it restores the documented
behavior that a goal will be invalidated as soon as one drops the handle.

Signed-off-by: Janosch Machowinski <J.Machowinski@cellumation.com>